removed warnings
extracted KNNBone from KNNSkeleton
removed namespance KNNRigger
