+++ Mocap-Loader/Assets/KNN Rigger/KNNBone.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b6a5112d01c55b4489ba1c6c31bfb00d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
