+++ Mocap-Loader/Assets/KNN Rigger/KNNRig.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using KNNRigger;
 using UnityEngine.Animations.Rigging;
 
 public class KNNRig : MonoBehaviour
